+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * [](#improved)
     * Improved `permissions` field
     * Added `field.show_label` and `field.label` display logic from frontend forms
+    * DEPRECATED `onAdminRegisterPermissions` event, use `RegisterPermissionsEvent::class` event instead
+    * DEPRECATED `Admin::setPermissions()` and `Admin::addPermissions()`, use `RegisterPermissionsEvent::class` event instead
+    * DEPRECATED `Admin::getPermissions()`, use `$grav['permissions']->getInstances()` instead
 1. [](#bugfix)
     * Fixed user profile when using `Flex Users` only in admin